refactor: 로그인관련 데이터 패칭 함수 리팩토링
